Create createNF.py
This srcipt creates the node features we defined from the files found in the given directory. (Therefore the path variable needs to be changed before running on the server or locally)
It combines flatten_and_features.py and stat_perf_sync.py to calculate node features, and then creates the defined structure of the node features. Finally it saves the results in .npy files. It does so in 50 match groups.